Bardiya: refactoring partition 2022/7/27

+++ algorithm/tourGenerators/FarthestInsertion.java
@@ -13,21 +13,21 @@ public class FarthestInsertion extends TourGenerator {
     @Override
     public Tour generateTour(Graph graph) {
         Vertex s = new Vertex(0), t = new Vertex(1);
-        for (Vertex v: graph.vertices)
-            for (Vertex u: graph.vertices)
-                if (!v.equals(u) && graph.getEdge(v,u).weight < graph.getEdge(s,t).weight) {
+        for (Vertex v : graph.vertices)
+            for (Vertex u : graph.vertices)
+                if (!v.equals(u) && graph.getEdge(v, u).weight < graph.getEdge(s, t).weight) {
                     s = v;
                     t = u;
                 }
         TreeSet<Edge> edges = new TreeSet<>();
-        edges.add(graph.getEdge(s,t));
+        edges.add(graph.getEdge(s, t));
         TreeSet<Vertex> selectedVertices = new TreeSet<>();
         selectedVertices.add(s);
         selectedVertices.add(t);
         TreeSet<Edge> remaining = new TreeSet<>();
-        for (Vertex vertex: graph.vertices)
+        for (Vertex vertex : graph.vertices)
             if (!vertex.equals(s) && !vertex.equals(t)) {
-                Edge es = graph.getEdge(s,vertex), et = graph.getEdge(t,vertex);
+                Edge es = graph.getEdge(s, vertex), et = graph.getEdge(t, vertex);
                 remaining.add(es.weight < et.weight ? es : et);
             }
         while (!remaining.isEmpty()) {
@@ -37,10 +37,10 @@ public class FarthestInsertion extends TourGenerator {
             List<Edge> deletions = new ArrayList<>();
             List<Edge> additions = new ArrayList<>();
             remaining.remove(e);
-            for (Edge edge: remaining) {
+            for (Edge edge : remaining) {
                 Vertex u = selectedVertices.contains(edge.v) ? edge.u : edge.v;
-                Edge newEdge = graph.getEdge(u,v);
-                if (edge.weight > newEdge.weight){
+                Edge newEdge = graph.getEdge(u, v);
+                if (edge.weight > newEdge.weight) {
                     deletions.add(edge);
                     additions.add(newEdge);
                 }
@@ -48,10 +48,10 @@ public class FarthestInsertion extends TourGenerator {
             deletions.forEach(remaining::remove);
             remaining.addAll(additions);
             Edge min = edges.first();
-            for (Edge edge: edges)
+            for (Edge edge : edges)
                 if (insertionCost(edge, v) < insertionCost(min, v))
                     min = edge;
-            if(edges.size() > 1)
+            if (edges.size() > 1)
                 edges.remove(min);
             edges.add(graph.getEdge(min.v, v));
             edges.add(graph.getEdge(min.u, v));
@@ -59,7 +59,7 @@ public class FarthestInsertion extends TourGenerator {
         return new Tour(new ArrayList<>(edges));
     }
 
-    private int insertionCost(Edge edge, Vertex v){
+    private double insertionCost(Edge edge, Vertex v) {
         return graph.getEdge(edge.v, v).weight + graph.getEdge(edge.u, v).weight - edge.weight;
     }
 }
